requirements.txt
streamlit
sacrebleu
python-Levenshtein
pandas
matplotlib
seaborn
